refactor: move the feature-flag dialog loader to the feature root
ESLint declares `@/features/*/components/**` private, and with the barrel gone
the loader is this feature's only public entry point — so it cannot live in the
folder the rule forbids importing from. Moving it to the feature root (a layout
feature-architecture.md sanctions) makes the import compliant and lets the
eslint-disable in SafeSidebarContent go away.
